content/response_integrations/google/anomali_staxx/pyproject.toml
[project]
name = "AnomaliStaxx"
version = "6.0"
description = "Anomali STAXX provides bi-directional sharing of threat intelligence from STIX/TAXII sources that are in the cloud (such as Anomali Limo, http://hailataxii.com, an ISAC, or Anomali ThreatStream) or on premise. With Anomali STAXX, you can connect to STIX/TAXII servers, discover and configure their threat feeds, and poll (download) threat intelligence from those feeds. You can also import threat intelligence into Anomali STAXX and push (upload) selected observables to other STIX/TAXII servers."
requires-python = ">=3.11,<3.12"
dependencies = [ "environmentcommon", "python-dateutil>=2.9.0.post0", "requests==2.32.4", "tipcommon",]

[dependency-groups]
dev = [ "pytest>=9.0.3", "pytest-json-report>=1.5.0", "soar-sdk",]

[tool.uv]
[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv.sources.tipcommon]
path = "../../../../packages/tipcommon/whls/TIPCommon-1.0.12-py2.py3-none-any.whl"

[tool.uv.sources.environmentcommon]
path = "../../../../packages/envcommon/whls/EnvironmentCommon-1.0.1-py2.py3-none-any.whl"

[tool.uv.sources.soar-sdk]
git = "https://github.com/chronicle/soar-sdk.git"
